Split up invoker tests
